requirements.txt
Flask==2.0.2
Werkzeug==2.0.2
Pillow==8.4.0
numpy==1.25.2
tensorflow==2.15.0
keras==2.15.0
